Oops, must use 64-bit integers to hold a Unix time multiplied by 10 or 100.
git-svn-id: http://svn/svn/hcr/trunk@1696 2e3f476d-dfe7-4101-bf40-6ec35b1080dd

+++ src/cmigitsDaemon/Ts2CmigitsFmqThread.cpp
@@ -141,11 +141,11 @@ Ts2CmigitsFmqThread::_doNextPulse() {
     // Test if the integer seconds, tenths of seconds, or hundredths of seconds
     // changed from the last pulse.
     bool secondsChanged = 
-            (int(_prevPulseDataTime) % 10) != (int(pulseDataTime) % 10);
+            (int64_t(_prevPulseDataTime) % 10) != (int64_t(pulseDataTime) % 10);
     bool tenthsChanged =
-            (int(_prevPulseDataTime * 10) % 10) != (int(pulseDataTime * 10) % 10);
+            (int64_t(_prevPulseDataTime * 10) % 10) != (int64_t(pulseDataTime * 10) % 10);
     bool hundredthsChanged =
-            (int(_prevPulseDataTime * 100) % 10) != (int(pulseDataTime * 100) % 10);
+            (int64_t(_prevPulseDataTime * 100) % 10) != (int64_t(pulseDataTime * 100) % 10);
 
     // Keep the processing time for this pulse and its data time
     _prevPulseProcessTime = pulseProcessTime;
